ATMO index data

get ATMO index data as a data.frame. The data is only containing the last year of daily historic.

read_atmo()

Value

a data.frame with the following columns:

date_time: (POSIXct with time zone set to Europe/Paris), date-time of index

score: (integer), air quality index

Details

The air quality index ranges from 1 (very good) to 10 (very bad). It characterizes in a simple and comprehensive way the air quality of an urban agglomeration.

It is composed of 4 sub index (all ranging from 1 to 10), each being representative of an air pollutant:

fine particles (PM10)

ozone (O3)

nitrogen dioxide (NO2)

sulfur dioxide (SO2)

The highest sub-index of the 4 presented above will be the index of the day.

It is not possible to reproduce exact calculations as so2 is not available through the web service or with read_measures().

See also

Examples

if( is_magellan_available() ){
  read_atmo()
}

#>      date_time score
#> 1   2018-08-30     3
#> 2   2018-08-31     4
#> 3   2018-09-01     4
#> 4   2018-09-02     5
#> 5   2018-09-03     4
#> 6   2018-09-04     4
#> 7   2018-09-05     5
#> 8   2018-09-06     4
#> 9   2018-09-07     4
#> 10  2018-09-08     4
#> 11  2018-09-09     5
#> 12  2018-09-10     4
#> 13  2018-09-11     4
#> 14  2018-09-12     4
#> 15  2018-09-13     4
#> 16  2018-09-14     4
#> 17  2018-09-15     3
#> 18  2018-09-16     4
#> 19  2018-09-17     5
#> 20  2018-09-18     4
#> 21  2018-09-19     4
#> 22  2018-09-20     4
#> 23  2018-09-21     3
#> 24  2018-09-22     3
#> 25  2018-09-23     2
#> 26  2018-09-24     3
#> 27  2018-09-25     3
#> 28  2018-09-26     4
#> 29  2018-09-27     6
#> 30  2018-09-28     5
#> 31  2018-09-29     3
#> 32  2018-09-30     3
#> 33  2018-10-01     3
#> 34  2018-10-02     3
#> 35  2018-10-03     3
#> 36  2018-10-04     4
#> 37  2018-10-05     5
#> 38  2018-10-06     4
#> 39  2018-10-07     3
#> 40  2018-10-08     4
#> 41  2018-10-09     5
#> 42  2018-10-10     5
#> 43  2018-10-11     3
#> 44  2018-10-12     3
#> 45  2018-10-13     3
#> 46  2018-10-14     4
#> 47  2018-10-15     4
#> 48  2018-10-16     5
#> 49  2018-10-17     4
#> 50  2018-10-18     5
#> 51  2018-10-19     6
#> 52  2018-10-20     5
#> 53  2018-10-21     6
#> 54  2018-10-22     4
#> 55  2018-10-23     4
#> 56  2018-10-24     4
#> 57  2018-10-25     5
#> 58  2018-10-26     3
#> 59  2018-10-27     3
#> 60  2018-10-28     3
#> 61  2018-10-29     3
#> 62  2018-10-30     3
#> 63  2018-10-31     3
#> 64  2018-11-01     3
#> 65  2018-11-02     3
#> 66  2018-11-03     5
#> 67  2018-11-04     5
#> 68  2018-11-05     4
#> 69  2018-11-06     3
#> 70  2018-11-07     2
#> 71  2018-11-08     3
#> 72  2018-11-09     3
#> 73  2018-11-10     3
#> 74  2018-11-11     3
#> 75  2018-11-12     3
#> 76  2018-11-13     3
#> 77  2018-11-14     4
#> 78  2018-11-15     3
#> 79  2018-11-16     3
#> 80  2018-11-17     4
#> 81  2018-11-18     3
#> 82  2018-11-19     3
#> 83  2018-11-20     4
#> 84  2018-11-21     6
#> 85  2018-11-22     5
#> 86  2018-11-23     5
#> 87  2018-11-24     3
#> 88  2018-11-25     3
#> 89  2018-11-26     4
#> 90  2018-11-27     4
#> 91  2018-11-28     2
#> 92  2018-11-29     3
#> 93  2018-11-30     4
#> 94  2018-12-01     2
#> 95  2018-12-02     3
#> 96  2018-12-03     3
#> 97  2018-12-04     3
#> 98  2018-12-05     3
#> 99  2018-12-06     3
#> 100 2018-12-07     3
#> 101 2018-12-08     3
#> 102 2018-12-09     4
#> 103 2018-12-10     3
#> 104 2018-12-11     5
#> 105 2018-12-12     4
#> 106 2018-12-13     3
#> 107 2018-12-14     4
#> 108 2018-12-15     4
#> 109 2018-12-16     2
#> 110 2018-12-17     3
#> 111 2018-12-18     3
#> 112 2018-12-19     3
#> 113 2018-12-20     3
#> 114 2018-12-21     3
#> 115 2018-12-22     3
#> 116 2018-12-23     2
#> 117 2018-12-24     3
#> 118 2018-12-25     4
#> 119 2018-12-26     6
#> 120 2018-12-27     5
#> 121 2018-12-28     6
#> 122 2018-12-29     6
#> 123 2018-12-30     3
#> 124 2018-12-31     3
#> 125 2019-01-01     3
#> 126 2019-01-02     3
#> 127 2019-01-03     3
#> 128 2019-01-04     4
#> 129 2019-01-05     4
#> 130 2019-01-06     5
#> 131 2019-01-07     4
#> 132 2019-01-08     3
#> 133 2019-01-09     3
#> 134 2019-01-10     3
#> 135 2019-01-11     3
#> 136 2019-01-12     3
#> 137 2019-01-13     3
#> 138 2019-01-14     3
#> 139 2019-01-15     3
#> 140 2019-01-16     4
#> 141 2019-01-17     3
#> 142 2019-01-18     5
#> 143 2019-01-19     4
#> 144 2019-01-20     4
#> 145 2019-01-21     8
#> 146 2019-01-22     5
#> 147 2019-01-23     3
#> 148 2019-01-24     6
#> 149 2019-01-25     5
#> 150 2019-01-26     3
#> 151 2019-01-27     3
#> 152 2019-01-28     3
#> 153 2019-01-29     3
#> 154 2019-01-30     3
#> 155 2019-01-31     3
#> 156 2019-02-01     3
#> 157 2019-02-02     3
#> 158 2019-02-03     3
#> 159 2019-02-04     3
#> 160 2019-02-05     3
#> 161 2019-02-06     3
#> 162 2019-02-07     3
#> 163 2019-02-08     3
#> 164 2019-02-09     3
#> 165 2019-02-10     4
#> 166 2019-02-11     3
#> 167 2019-02-12     6
#> 168 2019-02-13     6
#> 169 2019-02-14     6
#> 170 2019-02-15     7
#> 171 2019-02-16     8
#> 172 2019-02-17     5
#> 173 2019-02-18     5
#> 174 2019-02-19     4
#> 175 2019-02-20     7
#> 176 2019-02-21     8
#> 177 2019-02-22     8
#> 178 2019-02-23     4
#> 179 2019-02-24     4
#> 180 2019-02-25     5
#> 181 2019-02-26     7
#> 182 2019-02-27     8
#> 183 2019-02-28     5
#> 184 2019-03-01     4
#> 185 2019-03-02     3
#> 186 2019-03-03     4
#> 187 2019-03-04     4
#> 188 2019-03-05     3
#> 189 2019-03-06     3
#> 190 2019-03-07     4
#> 191 2019-03-08     4
#> 192 2019-03-09     3
#> 193 2019-03-10     4
#> 194 2019-03-11     4
#> 195 2019-03-12     4
#> 196 2019-03-13     4
#> 197 2019-03-14     3
#> 198 2019-03-15     3
#> 199 2019-03-16     4
#> 200 2019-03-17     4
#> 201 2019-03-18     3
#> 202 2019-03-19     4
#> 203 2019-03-20     6
#> 204 2019-03-21     5
#> 205 2019-03-22     6
#> 206 2019-03-23     6
#> 207 2019-03-24     6
#> 208 2019-03-25     4
#> 209 2019-03-26     4
#> 210 2019-03-27     5
#> 211 2019-03-28     6
#> 212 2019-03-29     5
#> 213 2019-03-30     5
#> 214 2019-03-31     5
#> 215 2019-04-01     6
#> 216 2019-04-02     4
#> 217 2019-04-03     4
#> 218 2019-04-04     4
#> 219 2019-04-05     4
#> 220 2019-04-06     4
#> 221 2019-04-07     4
#> 222 2019-04-08     5
#> 223 2019-04-09     4
#> 224 2019-04-10     6
#> 225 2019-04-11     5
#> 226 2019-04-12     5
#> 227 2019-04-13     4
#> 228 2019-04-14     4
#> 229 2019-04-15     5
#> 230 2019-04-16     4
#> 231 2019-04-17     4
#> 232 2019-04-18     5
#> 233 2019-04-19     5
#> 234 2019-04-20     6
#> 235 2019-04-21     6
#> 236 2019-04-22     5
#> 237 2019-04-23     5
#> 238 2019-04-24     4
#> 239 2019-04-25     4
#> 240 2019-04-26     4
#> 241 2019-04-27     4
#> 242 2019-04-28     3
#> 243 2019-04-29     4
#> 244 2019-04-30     6
#> 245 2019-05-01     6
#> 246 2019-05-02     5
#> 247 2019-05-03     3
#> 248 2019-05-04     3
#> 249 2019-05-05     4
#> 250 2019-05-06     4
#> 251 2019-05-07     4
#> 252 2019-05-08     4
#> 253 2019-05-09     3
#> 254 2019-05-10     4
#> 255 2019-05-11     4
#> 256 2019-05-12     4
#> 257 2019-05-13     4
#> 258 2019-05-14     4
#> 259 2019-05-15     4
#> 260 2019-05-16     4
#> 261 2019-05-17     4
#> 262 2019-05-18     4
#> 263 2019-05-19     4
#> 264 2019-05-20     4
#> 265 2019-05-21     4
#> 266 2019-05-22     5
#> 267 2019-05-23     6
#> 268 2019-05-24     5
#> 269 2019-05-25     4
#> 270 2019-05-26     3
#> 271 2019-05-27     4
#> 272 2019-05-28     3
#> 273 2019-05-29     4
#> 274 2019-05-30     2
#> 275 2019-05-31     4
#> 276 2019-06-01     4
#> 277 2019-06-02     5
#> 278 2019-06-03     3
#> 279 2019-06-04     4
#> 280 2019-06-05     3
#> 281 2019-06-06     4
#> 282 2019-06-07     4
#> 283 2019-06-08     3
#> 284 2019-06-09     4
#> 285 2019-06-10     3
#> 286 2019-06-11     4
#> 287 2019-06-12     3
#> 288 2019-06-13     3
#> 289 2019-06-14     4
#> 290 2019-06-15     4
#> 291 2019-06-16     3
#> 292 2019-06-17     4
#> 293 2019-06-18     5
#> 294 2019-06-19     4
#> 295 2019-06-20     3
#> 296 2019-06-21     4
#> 297 2019-06-22     5
#> 298 2019-06-23     4
#> 299 2019-06-24     5
#> 300 2019-06-25     5
#> 301 2019-06-26     6
#> 302 2019-06-27     5
#> 303 2019-06-28     6
#> 304 2019-06-29     7
#> 305 2019-06-30     5
#> 306 2019-07-01     4
#> 307 2019-07-02     4
#> 308 2019-07-03     5
#> 309 2019-07-04     5
#> 310 2019-07-05     5
#> 311 2019-07-06     7
#> 312 2019-07-07     5
#> 313 2019-07-08     4
#> 314 2019-07-09     4
#> 315 2019-07-10     5
#> 316 2019-07-11     5
#> 317 2019-07-12     4
#> 318 2019-07-13     4
#> 319 2019-07-14     3
#> 320 2019-07-15     4
#> 321 2019-07-16     5
#> 322 2019-07-17     6
#> 323 2019-07-18     4
#> 324 2019-07-19     4
#> 325 2019-07-20     3
#> 326 2019-07-21     3
#> 327 2019-07-22     6
#> 328 2019-07-23     6
#> 329 2019-07-24     7
#> 330 2019-07-25     7
#> 331 2019-07-26     5
#> 332 2019-07-27     3
#> 333 2019-07-28     3
#> 334 2019-07-29     5
#> 335 2019-07-30     3
#> 336 2019-07-31     3
#> 337 2019-08-01     4
#> 338 2019-08-02     4
#> 339 2019-08-03     4
#> 340 2019-08-04     5
#> 341 2019-08-05     3
#> 342 2019-08-06     2
#> 343 2019-08-07     3
#> 344 2019-08-08     4
#> 345 2019-08-09     3
#> 346 2019-08-10     4
#> 347 2019-08-11     3
#> 348 2019-08-12     3
#> 349 2019-08-13     3
#> 350 2019-08-14     3
#> 351 2019-08-15     3
#> 352 2019-08-16     3
#> 353 2019-08-17     2
#> 354 2019-08-18     3
#> 355 2019-08-19     3
#> 356 2019-08-20     4
#> 357 2019-08-21     4
#> 358 2019-08-22     5
#> 359 2019-08-23     6
#> 360 2019-08-24     5
#> 361 2019-08-25     7
#> 362 2019-08-26     6
#> 363 2019-08-27     7
#> 364 2019-08-28     5
#> 365 2019-08-29     4